Simple model example (3 layers of cnn/maxpooling)

+++ cifar.py
@@ -26,8 +26,8 @@ def load_dataset():
     metadata = unpickle_file('batches.meta')
     label_names, batch_size = metadata[b'label_names'], metadata[b'num_cases_per_batch']
     num_images = NUM_BATCHES * batch_size
-    images = np.zeros(shape=[num_images, IMAGE_SIZE, IMAGE_SIZE, NUM_CHANNELS], dtype=np.float)
-    labels = np.zeros(shape=[num_images], dtype=np.int)
+    images = np.zeros(shape=[num_images, IMAGE_SIZE, IMAGE_SIZE, NUM_CHANNELS], dtype=np.float32)
+    labels = np.zeros(shape=[num_images], dtype=np.int32)
     for batch_index in range(NUM_BATCHES):
         batch = unpickle_file('data_batch_{}'.format(batch_index + 1))
         begin_i, end_i = batch_index * batch_size, (batch_index + 1) * batch_size
